pyproject.toml
[project]
name = "video-classifier"
version = "0.1.0"
description = "Zero-shot video classifier using CLIP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.40.0",
    "opencv-python>=4.9.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
classify = "video_classifier.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["video_classifier"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
